#1881 reorganized the views
