+++ view/statistics-files-certificates.js
@@ -27,10 +27,10 @@ var _                 = require('mano').i18n.bind('View: Statistics')
   , dateFromToBlock         = require('./components/filter-bar/select-date-range-safe-fallback')
   , getDynamicUrl           = require('./utils/get-dynamic-url');
 
-exports._parent        = require('./statistics-flow');
+exports._parent        = require('./statistics-files');
 exports._customFilters = Function.prototype;
 
-exports['flow-nav']                = { class: { 'submitted-menu-item-active': true } };
+exports['files-nav']                = { class: { 'submitted-menu-item-active': true } };
 exports['flow-by-certificate-nav'] = { class: { 'pills-nav-active': true } };
 
 var serviceToCertLegacyMatch = { '': [] };
@@ -53,12 +53,12 @@ db.BusinessProcess.extensions.forEach(function (ServiceType) {
 
 exports['statistics-main'] = function () {
 	var queryHandler, data = new ObservableValue({})
-	  , pagination = new Pagination('/flow/'), handlerConf, params;
+	  , pagination = new Pagination('/files/'), handlerConf, params;
 
 	handlerConf = queryHandlerConf.slice(0);
 	handlerConf.push(pageQuery, serviceQuery, certificateQuery);
 	queryHandler = setupQueryHandler(handlerConf,
-		location, '/flow/');
+		location, '/files/');
 
 	params = queryHandler._handlers.map(function (handler) {
 		return handler.name;
@@ -113,7 +113,7 @@ exports['statistics-main'] = function () {
 		});
 	});
 
-	div({ class: 'block-pull-up' }, form({ action: '/flow/', autoSubmit: true },
+	div({ class: 'block-pull-up' }, form({ action: '/files/', autoSubmit: true },
 		section({ class: 'date-period-selector-positioned-on-submenu' }, dateFromToBlock()),
 		section({ class: 'section-primary users-table-filter-bar display-flex flex-wrap' },
 			div(
